polybar/playerctl: In-script loop, signals support
Module now fully works on a built-in loop and can handle signals, making
it possible (and actually making it) to add a switcher for the extra
modes.

One of the first extra modes allow to hide detailed metadata and show
only simplified output in format "{icon} Playback". Can be activated on
RMB click.