polybar/playerctl: Freeze without termination
Now script will indefinitely run in the background if no required
utility is detected.

Previous solution was to just error-exit the script, but this will
trigger the polybar to restart it after some time (declared by delay
configuration). There is no way to make polybar stop the module startup
on error exit code, so I gave up on it and made it myself.